Type of Segue
Segue about iPhone : push, modal, custom .
push : need a UINavigationViewController;
modal : cover the presenting view controller, need protocol to get
back.. A simple way is just add a delegate method ( let the presenting
view controller as the delegate object , and then implements the
delegate method )
custom : hate it , dislike it . with the demo , I use the storyboard
segue , which I cover the perform method ,and write a lots code , and
the custom view controller is the destination view controller. Of
course if I want to come back to my source view controller , I also
need the protocol ,just same as the model delegate.